[test] redistribute edge-case tests into unit files, remove duplicates
Move tests from edge-cases.test.ts into bondRiskFee, bidTooLowPenalty,
constraints, eligibility, sam, calculations, and updatePaidUndelegation
test files. Remove duplicate sam effective-bid test. 100 tests, 11 suites.
